+++ tests/python/test_grammar_compiler.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel
 from transformers import AutoTokenizer
 
 from xgrammar import BuiltinGrammar, CompiledGrammar, GrammarMatcher, TokenizerInfo
-from xgrammar.xgrammar import CachedGrammarCompiler
+from xgrammar.matcher import GrammarCompiler
 
 
 def test_compiled_grammar():
@@ -41,11 +41,11 @@ def test_compiled_grammar():
     check_matcher(matcher_2)
 
 
-def test_cached_grammar_compiler_json():
+def test_grammar_compiler_json():
     tokenizer = AutoTokenizer.from_pretrained("meta-llama/Llama-2-7b-chat-hf")
     tokenizer_info = TokenizerInfo.from_huggingface(tokenizer)
     time_start = time.monotonic_ns()
-    cached_grammar_compiler = CachedGrammarCompiler(tokenizer_info)
+    grammar_compiler = GrammarCompiler(tokenizer_info)
     time_end = time.monotonic_ns()
     print(f"Time to init cached grammar compiler: {(time_end - time_start) / 1e3} us")
 
@@ -57,33 +57,33 @@ def test_cached_grammar_compiler_json():
         assert matcher.is_terminated()
 
     time_start = time.monotonic_ns()
-    compiled_grammar = cached_grammar_compiler.compile_json_grammar()
+    compiled_grammar = grammar_compiler.compile_builtin_json_grammar()
     time_end = time.monotonic_ns()
     print(f"Time to get compiled grammar: {(time_end - time_start) / 1e3} us")
     matcher = GrammarMatcher(compiled_grammar, terminate_without_stop_token=True)
     check_matcher(matcher)
 
     time_start = time.monotonic_ns()
-    compiled_grammar = cached_grammar_compiler.compile_json_grammar()
+    compiled_grammar = grammar_compiler.compile_builtin_json_grammar()
     time_end = time.monotonic_ns()
     print(f"Time to get compiled grammar again: {(time_end - time_start) / 1e3} us")
     matcher = GrammarMatcher(compiled_grammar, terminate_without_stop_token=True)
     check_matcher(matcher)
 
-    cached_grammar_compiler.clear()
+    grammar_compiler.clear()
 
     time_start = time.monotonic_ns()
-    compiled_grammar = cached_grammar_compiler.compile_json_grammar()
+    compiled_grammar = grammar_compiler.compile_builtin_json_grammar()
     time_end = time.monotonic_ns()
     print(f"Time to get compiled grammar after clear: {(time_end - time_start) / 1e3} us")
     matcher = GrammarMatcher(compiled_grammar, terminate_without_stop_token=True)
     check_matcher(matcher)
 
 
-def test_cached_grammar_compiler_json_schema():
+def test_grammar_compiler_json_schema():
     tokenizer = AutoTokenizer.from_pretrained("meta-llama/Llama-2-7b-chat-hf")
     tokenizer_info = TokenizerInfo.from_huggingface(tokenizer)
-    cached_grammar_compiler = CachedGrammarCompiler(tokenizer_info)
+    grammar_compiler = GrammarCompiler(tokenizer_info)
 
     class MainModel(BaseModel):
         integer_field: int
@@ -110,7 +110,7 @@ def test_cached_grammar_compiler_json_schema():
         instance_str = instance.model_dump_json(indent=indent, round_trip=True)
 
         time_start = time.monotonic_ns()
-        compiled_grammar = cached_grammar_compiler.compile_json_schema_grammar(
+        compiled_grammar = grammar_compiler.compile_json_schema(
             MainModel, indent=indent, separators=separators
         )
         time_end = time.monotonic_ns()
@@ -127,7 +127,7 @@ def test_cached_grammar_compiler_json_schema():
     check_with_fmt(2, None, "3")
     check_with_fmt(2, (",", ": "), "4")
 
-    cached_grammar_compiler.clear()
+    grammar_compiler.clear()
 
     check_with_fmt(None, (",", ":"), "5")
 
@@ -176,10 +176,10 @@ schema_instances = [
 ]
 
 
-def test_cached_grammar_compiler_json_schema_concurrent():
+def test_grammar_compiler_json_schema_concurrent():
     tokenizer = AutoTokenizer.from_pretrained("meta-llama/Llama-2-7b-chat-hf")
     tokenizer_info = TokenizerInfo.from_huggingface(tokenizer)
-    cached_grammar_compiler = CachedGrammarCompiler(tokenizer_info)
+    grammar_compiler = GrammarCompiler(tokenizer_info)
 
     def check_matcher(matcher: GrammarMatcher, instance_str: str):
         assert matcher.vocab_size == 32000
@@ -195,7 +195,7 @@ def test_cached_grammar_compiler_json_schema_concurrent():
         schema_id = id % num_schemas
         time_mid = time.monotonic_ns()
         print(f"Thread {id} start compile grammar {schema_id}: {(time_mid - time_start) / 1e3} us")
-        compiled_grammar = cached_grammar_compiler.compile_json_schema_grammar(
+        compiled_grammar = grammar_compiler.compile_json_schema(
             schema, indent=None, separators=(",", ":"), strict_mode=True
         )
         time_end = time.monotonic_ns()
